requirements.txt
Flask
numpy
pandas
scikit-surprise